Modify styling and rendering of the add menu
Changed each of the "add-something" components so that having a button is optional.
By default, the button still renders. However, if the component is called with
"hideButton=true", the dialog box will render on its own without needing to click a button.
So in add-knowledge-menu, each component is conditionally rendered when selected,
with hideButton activated.